+++ LLD-Practice-Questions/src/hotel_booking_system/hotelservice/HotelService.java
@@ -1,7 +1,9 @@
 package hotel_booking_system.hotelservice;
 
+import java.math.BigDecimal;
+
 public interface HotelService {
-    Double getPrice();
+    BigDecimal getPrice();
     String getDescription();
     String getName();
 }
