+++ LLD-Practice-Questions/src/hotel_booking_system/hotelservice/SingleService.java
@@ -1,18 +1,20 @@
 package hotel_booking_system.hotelservice;
 
+import java.math.BigDecimal;
+
 public class SingleService implements HotelService {
     private final String name;
-    private final Double price;
+    private final BigDecimal price;
     private final String description;
 
-    public SingleService(String name, Double price, String description) {
+    public SingleService(String name, BigDecimal price, String description) {
         this.name = name;
         this.price = price;
         this.description = description;
     }
 
     @Override
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
